requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

pydantic[email]>=2.3.0

bcrypt>=4.0.0
PyJWT>=2.8.0

python-dotenv>=1.0.0
requests>=2.31.0